+++ tests/test_json_schema.py
@@ -9,15 +9,6 @@ from statham.dsl.parser import parse_element
 from statham.dsl.exceptions import FeatureNotImplementedError, ValidationError
 from tests.helpers import no_raise
 
-# This is a magic pytest constant.
-# pylint: disable=invalid-name
-pytestmark = [
-    pytest.mark.skipif(
-        os.getenv("OFFICIAL_TEST_SUITE", "false").lower() not in ("true", "1"),
-        reason=("Test against JSONSchema official test pack."),
-    )
-]
-
 
 NOT_IMPLEMENTED = (
     "optional",